Refactor TomographModel to put parameter handling in a superclass. Perform some code cleanup.
