+++ experiments/cvpr-2024/vcd_figs_for_abst.py
@@ -63,6 +63,7 @@ if __name__ == "__main__":
     # Set reconstruction parameter values
     parallel_model.set_params(sharpness=sharpness, verbose=1)
     parallel_model.set_params(partition_sequence=[0, 1, 2, 3, 1, 2, 3, 2, 3, 3, 0, 1, 2, 3, 1, 2, 3, 2, 3, 3])
+    granularity = np.array(parallel_model.get_params('granularity'))
 
     # Print out model parameters
     parallel_model.print_params()
@@ -70,22 +71,20 @@ if __name__ == "__main__":
     # ##########################
     # Perform VCD reconstruction
     recon_vcd, fm_rmse_vcd = parallel_model.recon(sinogram, weights=weights)
-    granularity = np.array(parallel_model.get_params('granularity'))
-    partition_sequence = parallel_model.gen_partition_sequence(num_iterations=13)
+    default_partition_sequence = parallel_model.get_params('partition_sequence')
+    partition_sequence = mbirjax.gen_partition_sequence(default_partition_sequence, num_iterations=13)
     granularity_sequence_vcd = granularity[partition_sequence]
 
     # Perform GD reconstruction
     parallel_model.set_params(partition_sequence=[0,])
     recon_gd, fm_rmse_gd = parallel_model.recon(sinogram, weights=weights)
-    granularity = np.array(parallel_model.get_params('granularity'))
-    partition_sequence = parallel_model.gen_partition_sequence(num_iterations=13)
+    partition_sequence = mbirjax.gen_partition_sequence(partition_sequence, num_iterations=13)
     granularity_sequence_gd = granularity[partition_sequence]
 
     # Perform CD reconstruction
     parallel_model.set_params(partition_sequence=[3,])
     recon_cd, fm_rmse_cd = parallel_model.recon(sinogram, weights=weights)
-    granularity = np.array(parallel_model.get_params('granularity'))
-    partition_sequence = parallel_model.gen_partition_sequence(num_iterations=13)
+    partition_sequence = mbirjax.gen_partition_sequence(partition_sequence, num_iterations=13)
     granularity_sequence_cd = granularity[partition_sequence]
     # ##########################
