+++ mbirjax/__init__.py
@@ -1,4 +1,5 @@
 from .projectors import *
+from .parameter_handler import *
 from .tomography_model import *
 from .parallel_beam import *
 from .cone_beam import *
